+++ example/lib/tab/tv_tab_bar_foundation_sample.dart
@@ -125,7 +125,7 @@ final class _TvTabBarFoundationSampleState
                     focusScopeNode: _focusScopeNode,
                     mainAxisSize: MainAxisSize.min,
                     mainAxisAlignment: MainAxisAlignment.start,
-                    spacing: 16,
+                    separatorBuilder: (_, _) => const SizedBox(width: 16),
                     onDown: (_, _, isOutOfScope) {
                       if (isOutOfScope) {
                         _contentFocusNode.requestFocus();
@@ -143,8 +143,8 @@ final class _TvTabBarFoundationSampleState
                           key: TvTabBarFoundationSample.tabsKeys[i],
                           index: i,
                           currentIndex: _currentIndex,
-                          onFocusChanged: (node) {
-                            if (node.hasFocus) {
+                          onFocusChanged: (_, hasFocus) {
+                            if (hasFocus) {
                               _updateSelectionConstraints();
                             }
                           },
@@ -313,7 +313,7 @@ final class TabItem extends StatelessWidget {
 
   final int index;
   final int currentIndex;
-  final void Function(FocusNode)? onFocusChanged;
+  final void Function(FocusNode, bool)? onFocusChanged;
 
   @override
   Widget build(BuildContext context) {