+++ lib/src/foundation/navigation/menu/menu.dart
@@ -1,6 +1,6 @@
 library;
 
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter/services.dart';
 import 'package:tv_plus/src/foundation/foundation.dart';