+++ lib/src/foundation/scroll/tv_grid_view.dart
@@ -120,7 +120,7 @@ final class TvGridView extends BoxScrollView {
       onRight: onRight,
       onFocusChanged: onFocusChanged,
       onFocusDisabledWhenWasFocused: onFocusDisabledWhenWasFocused,
-      builder: (_) => super.build(context),
+      builder: (context, _) => super.build(context),
     );
   }
 }